+++ config/hard/Boggle/code/CGCode.js
@@ -1,0 +1,13 @@
+const line1 = readline();
+const line2 = readline();
+const line3 = readline();
+const line4 = readline();
+const n = parseInt(readline());
+for (let i = 0; i < n; i++) {
+    const w = readline();
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');