+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "Highest truncated pyramid".
 - configuration for "Sign (programming) language".
 - configuration for "Secret Message Decoding".
+- configuration for "Boggle".
 
 ## [1.12.0] - 2025-02-28
 ### Added